rename index file
