+++ apps/web/src/utils/proofs/sybil_resistance.ts
@@ -1,5 +1,5 @@
 import { getAttestations } from '@coinbase/onchainkit/identity';
-import { kv } from 'apps/web/src/utils/kv/kv';
+import { kv } from 'apps/web/src/utils/kv';
 import { CoinbaseProofResponse } from 'apps/web/pages/api/proofs/coinbase';
 import RegistrarControllerABI from 'apps/web/src/abis/RegistrarControllerABI';
 import {
